+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## v2.0.7
+* Added doob/support - Gives link to the Doob support server.
+* Added doob/invite - Gives link to invite Doob to another server.
 
 ## v2.0.6
 * New Lucky Dog! (Kevin from [@Weesterner](https://twitter.com/weesterner)!)